+++ src/clj/withings_client/misc.clj
@@ -4,8 +4,8 @@
    [java-time :as jt]))
 
 (defn datetime->timestamp
-  "input: yyyy-MM-DD hh:mm:ss
-   returns timestamp(int)"
+  "input:  yyyy-MM-DD hh:mm:ss
+   return: timestamp (integer)"
   [s]
   (when (seq s)
     (let [[date time] (str/split s #" ")]
@@ -15,9 +15,21 @@
             1000))))
 
 (comment
-  (defn f [s]
-    (when (seq s)
-      s))
-  (f "abc")
-  (f nil)
+    (datetime->timestamp "2022-08-31 12:34:56")
+    (<  (datetime->timestamp "2022-08-31 12:34:56")
+        2000000000)
+  )
+
+(defn abbrev
+  "default length = 8 abbreviation.
+   if want other than 8 abbrev, say n,
+   use (abbrev n s)"
+  ([s]
+   (abbrev 8 s))
+  ([n s]
+   (let [re (re-pattern (format "^(.{%d}).*" n))]
+     (str/replace s re (str "$1" "...")))))
+
+(comment
+  (abbrev (-> (range 10) (str/join)))
   )